Points removal from ranked
Removed the points portion from the rank text when a jury ranks a country in their top 10. For posterity, last commit here as well. Deleted file and replaced with this newly named file. Added a country loop with a variable end point since a non qual country ranks all 26 countries where a qual country only ranks 25 and None for itself.